verificaciones de productos y proveedores

+++ app/compras/models.py
@@ -16,7 +16,7 @@ class Producto(models.Model):
 
     nombre = models.CharField(max_length=50)
     precio = models.DecimalField(max_digits=10, decimal_places=2)
-    stock_actual = models.DecimalField(max_digits=10, decimal_places=0)
+    stock_actual = models.IntegerField()
     proveedor = models.ForeignKey(Proveedor, on_delete=models.CASCADE)
 
     def __str__(self):
